+++ tests/CreatesApplication.php
@@ -1,16 +1,48 @@
 <?php namespace GeneaLabs\LaravelModelCaching\Tests;
 
 use GeneaLabs\LaravelModelCaching\Providers\Service as LaravelModelCachingService;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Author;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Book;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Profile;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Publisher;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Store;
 use Orchestra\Database\ConsoleServiceProvider;
 
 trait CreatesApplication
 {
+    protected $cache;
+
     public function setUp()
     {
         parent::setUp();
 
         $this->withFactories(__DIR__ . '/database/factories');
         $this->loadMigrationsFrom(__DIR__ . '/database/migrations');
+
+        $this->cache = cache()
+            ->store(config('laravel-model-caching.store'));
+
+        $this->cache()->flush();
+        $publishers = factory(Publisher::class, 10)->create();
+        factory(Author::class, 10)->create()
+            ->each(function ($author) use ($publishers) {
+                factory(Book::class, random_int(2, 10))->make()
+                    ->each(function ($book) use ($author, $publishers) {
+                        $book->author()->associate($author);
+                        $book->publisher()->associate($publishers[rand(0, 9)]);
+                        $book->save();
+                    });
+                factory(Profile::class)->make([
+                    'author_id' => $author->id,
+                ]);
+            });
+
+        $bookIds = (new Book)->all()->pluck('id');
+        factory(Store::class, 10)->create()
+            ->each(function ($store) use ($bookIds) {
+                $store->books()->sync(rand($bookIds->min(), $bookIds->max()));
+            });
+        $this->cache()->flush();
     }
 
     /**
@@ -22,5 +54,23 @@ trait CreatesApplication
             LaravelModelCachingService::class,
             ConsoleServiceProvider::class,
         ];
+    }
+
+    protected function getEnvironmentSetUp($app)
+    {
+        $app['config']->set('database.redis.default', [
+            'host' => env('REDIS_HOST', '192.168.10.10'),
+        ]);
+        $app['config']->set('database.redis.model-cache', [
+            'host' => env('REDIS_HOST', '192.168.10.10'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', 6379),
+            'database' => 1,
+        ]);
+        $app['config']->set('cache.stores.model', [
+            'driver' => 'redis',
+            'connection' => 'model-cache',
+        ]);
+        $app['config']->set('laravel-model-caching.store', 'model');
     }
 }