+++ tests/Unit/DisabledCachedBuilderTest.php
@@ -22,33 +22,6 @@ class DisabledCachedBuilderTest extends UnitTestCase
 {
     use RefreshDatabase;
 
-    public function setUp()
-    {
-        parent::setUp();
-
-        cache()->flush();
-        $publishers = factory(Publisher::class, 10)->create();
-        factory(Author::class, 10)->create()
-            ->each(function ($author) use ($publishers) {
-                factory(Book::class, random_int(2, 10))->make()
-                    ->each(function ($book) use ($author, $publishers) {
-                        $book->author()->associate($author);
-                        $book->publisher()->associate($publishers[rand(0, 9)]);
-                        $book->save();
-                    });
-                factory(Profile::class)->make([
-                    'author_id' => $author->id,
-                ]);
-            });
-
-        $bookIds = (new Book)->all()->pluck('id');
-        factory(Store::class, 10)->create()
-            ->each(function ($store) use ($bookIds) {
-                $store->books()->sync(rand($bookIds->min(), $bookIds->max()));
-            });
-        cache()->flush();
-    }
-
     public function testAvgModelResultsIsNotCached()
     {
         $authorId = (new Author)
@@ -62,7 +35,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResult = (new UncachedAuthor)
@@ -111,7 +84,8 @@ class DisabledCachedBuilderTest extends UnitTestCase
 
         for ($index = 0; $index < $cachedChunks['authors']->count(); $index++) {
             $key = $cachedChunks['keys'][$index];
-            $cachedResults = cache()->tags($tags)
+            $cachedResults = $this->cache()
+                ->tags($tags)
                 ->get($key);
 
             $this->assertNull($cachedResults);
@@ -132,7 +106,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResults = cache()
+        $cachedResults = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResults = (new UncachedAuthor)
@@ -156,7 +130,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResults = cache()
+        $cachedResults = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResults = collect(
@@ -180,7 +154,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesauthor',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResult = (new UncachedAuthor)
@@ -203,7 +177,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResults = cache()
+        $cachedResults = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResults = (new UncachedAuthor)
@@ -227,7 +201,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResult = (new UncachedAuthor)
@@ -251,7 +225,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResult = (new UncachedAuthor)
@@ -275,7 +249,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResults = cache()
+        $cachedResults = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResults = (new UncachedAuthor)
@@ -299,7 +273,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);
         $liveResult = (new UncachedAuthor)
@@ -323,7 +297,7 @@ class DisabledCachedBuilderTest extends UnitTestCase
             'genealabslaravelmodelcachingtestsfixturesprofile',
         ];
 
-        $cachedResult = cache()
+        $cachedResult = $this->cache()
             ->tags($tags)
             ->get($key);